Port TMapMgr::BuildOrLoadGlobalMapStateForSession (0x50ec90, 98.0%) + SetMapRecordFlagA3AndPropagateToChildren (0x50f6b0, 100%)
- 0x50ec90 (874B, virtual slot 0xb): full spin-guarded map build/load pipeline
  (replay reload, scenario table load with bail-out, tuning-string generation via
  TMapMaker, tile post-passes, region-class flood assignment). Param 1 retyped
  from junk (CString, char*) to const char* mapStreamName (caller 0x511e80
  passes "mapdata"; only tested non-null).
- 0x50f6b0 (110B): recursive region-class flood, re-attributed from junk
  TSpaceCommand to TMapMgr; 100% match.
- Deferred bodies for 0x50f860 RebuildTileOwnerNeighborCachesAndFallbackAssignments
  and 0x518540 LoadScenarioMapStateFromTableResource(int) (noop baseline updated).
- Slot 0xd UpdateMapTileAdjacencyMasksAndVariantForTile param retyped uint->short
  (orig prologue does movsx bx; kills a double induction variable in both 0x1950
  loops of 0x50ec90). TGameSetupPicture caller loop var aligned to short.
- TMapMaker driver arg2 retyped int -> TGlobalMapCityScoreRecord* (cityScoreTable).
- Matching lessons: record-pointer locals as 'base + index' give the orig
  add-into-base form; named CString local lands in the dead arg slot and
  re-materializes its address with lea; int param (not short cast) for
  movsx+dec eax arg shape.

Verified: just build clean, gates pass (noop/stub-count baselines updated),
109 tooling tests OK, stats +11 improved / 4 sub-2pp wobble accepted.

+++ src/game/TGameSetupPicture.cpp
@@ -27,8 +27,7 @@ TGameSetupPicture::TGameSetupPicture() : TNoHilitePicture() {}
 TGameSetupPicture::~TGameSetupPicture() {}
 
 // FUNCTION: IMPERIALISM 0x005758e0
-void TGameSetupPicture::NoOpUiLifecycleHook(int arg) {
-}
+void TGameSetupPicture::NoOpUiLifecycleHook(int arg) {}
 
 // Main-menu button dispatcher. Only commandId 0x14/0x0a/0x22 (button-activation
 // codes) are handled; anything else forwards straight to the base class. The
@@ -97,10 +96,9 @@ void TGameSetupPicture::HandleEvent(int commandId, TEventHandler* sourceHandler,
       g_pSimMgr->RebuildGlobalOrderManagersAndCapabilityState(1);
       g_pGlobalMapState->AllocateAndResetTerrainAndCityScoreTables();
       g_pGlobalMapState->LoadPoliticalMapRegionSubtypeTableFromResourceStream();
-      for (unsigned int tileIndex = 0; tileIndex < 0x1950; ++tileIndex) {
+      for (short tileIndex = 0; tileIndex < 0x1950; ++tileIndex) {
         g_pGlobalMapState->UpdateMapTileAdjacencyMasksAndVariantForTile(tileIndex);
-        g_pGlobalMapState->UpdateTileNeighborBorderInfluenceCounters(
-            static_cast<short>(tileIndex), 0);
+        g_pGlobalMapState->UpdateTileNeighborBorderInfluenceCounters(tileIndex, 0);
       }
       postEventCode = 0x3c0;
     } else {
